Add option to disable all keyboard shortcuts
Introduces a new 'Disable Keyboard Shortcuts' setting in the configuration, UI, and backend. When enabled, all keyboard shortcuts except for opening the panel are disabled, and the shortcuts tab is hidden and deactivated in the panel. This provides users with the ability to globally turn off shortcut functionality if desired.

For #64